refactor(identity): split the revocation cascade out of oauth_grant.go
oauth_grant.go had grown to 513 lines (craft-static's 500-line ceiling)
after this branch's supersede/audit-only comments. The revocation
cascade (lockGrant onward) is a coherent concept of its own, already
read by the separate RFC 7009 handler in oauth_revoke.go — split it
into oauth_grantrevocation.go rather than trim the comments that
craft-static's own review just asked for.
